Extract outcome review API module
